+++ frontend/src/components/homepage/LoginSection.jsx
@@ -1,0 +1,18 @@
+export default function LoginSection({ handleUserSelect }) {
+  return (
+    <div className="flex flex-col items-center justify-center h-2/3 w-2/3 gap-4">
+      <h1 className="text-3xl text-letter font-jetbrains p-4">I'm a ...</h1>
+      <div className="flex flex-row gap-4 justify-center items-center">
+        <span className="text-primary text-6xl font-jetbrains">{`{`}</span>
+        <button className="main-btn" onClick={() => handleUserSelect(true)}>
+          teacher
+        </button>
+        <span className="text-letter text-5xl font-jetbrains">{`||`}</span>
+        <button className="main-btn" onClick={() => handleUserSelect(false)}>
+          student
+        </button>
+        <span className="text-primary text-6xl font-jetbrains">{`}`}</span>
+      </div>
+    </div>
+  );
+}
